% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCouncilStyle.R
\name{addPolygons}
\alias{addPolygons}
\title{Wrapper for addPolygons}
\usage{
addPolygons(map, smoothFactor = 0, weight = 0, fillOpacity = 1, ...)
}
\arguments{
\item{map}{A \code{leaflet} map}
}
\value{
A \code{leaflet} map with polygons added
}
\description{
All the same inputs apply as the leaflet::addPolygons function, we just use this wrapper to define the "defaults" for certain inputs
}
