% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCouncilStyle.R
\name{addSourceText}
\alias{addSourceText}
\title{Add a "Source" note to a leaflet that will be a static output}
\usage{
addSourceText(map, source_text, color = "#555555", fontSize = "15px", ...)
}
\arguments{
\item{map}{A \code{leaflet} map}

\item{source_text}{The text that you want added to the map}

\item{color}{}

\item{fontSize}{}
}
\value{
A \code{leaflet} map with a source note added in the bottom right for the councildown defined NYC frame
}
\description{
Add a "Source" note to a leaflet that will be a static output
}
