% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCouncilStyle.R
\name{colorBin}
\alias{colorBin}
\title{Wrapper for colorBin}
\usage{
colorBin(
  palette = "nycc_blue",
  domain = NULL,
  bins = 7,
  na.color = "#FFFFFF",
  ...
)
}
\arguments{
\item{domain}{Possible values to be mapped by \code{leaflet}}
}
\value{
A \code{leaflet} colorBin palette
}
\description{
All the same inputs apply as the leaflet::colorBin function, just use this wrapper to define the "defaults" for certain inputs
}
