% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/councilPopup.R
\name{councilPopup}
\alias{councilPopup}
\title{Basic Styling for Leaflet Popups}
\usage{
councilPopup(x)
}
\arguments{
\item{x}{The popup HTML}
}
\value{
A character vector
}
\description{
This function wraps its input in a \code{<div>} and sets council fonts for
body and headers.
}
\examples{

councilPopup("Test")
}
