% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_add_modified.R
\name{ggplot_add}
\alias{ggplot_add}
\title{Add custom objects to ggplot}
\usage{
ggplot_add(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add `object` to}

\item{object_name}{The name of the object to add}
}
\value{
A modified ggplot object
}
\description{
This generic allows you to add your own methods for adding custom objects to
a ggplot with [+.gg].
}
\keyword{internal}
