% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hook_pdfembed}
\alias{hook_pdfembed}
\title{Embed fonts in PDF images}
\usage{
hook_pdfembed(before, options, envir)
}
\arguments{
\item{before}{Is hook being run before chunk}

\item{options}{Chunk options}

\item{envir}{Environment}
}
\description{
This knitr chunk hook embeds fonts in PDF images so that the appear correctly
on all computers.
}
