% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nycc_cd_13}
\alias{nycc_cd_13}
\title{2013-2023 Council District Shapefile}
\format{
## `nycc_cd_13`
An sf collection with 51 rows and 7 columns w/ multiploygon geography:
\describe{
  \item{coun_dist}{Council District}
  \item{Shape_Leng, Shape_Area}{Length and Area of Council Dsitrict}
  \item{lab_coord}{Coordinates in Matrix Form}
  \item{lab_x lab_y}{Longitude and Latitude}
  \item{geography}{Multipolygon of Council District}
}
}
\source{
<https://s-media.nyc.gov/agencies/dcp/assets/files/zip/data-tools/bytes/nycc_22c.zip>
}
\usage{
nycc_cd_13
}
\description{
2013-2023 Council District Shapefile
}
\keyword{datasets}
