% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nycc_ct_10}
\alias{nycc_ct_10}
\title{2010 Census Tract Shapefile (clipped to shoreline)}
\format{
## `nycc_ct_10`
An sf collection with 2165 rows and 14 columns w/ multiploygon geography:
}
\source{
<https://s-media.nyc.gov/agencies/dcp/assets/files/zip/data-tools/bytes/nyct2010_23a.zip>
}
\usage{
nycc_ct_10
}
\description{
2010 Census Tract Shapefile (clipped to shoreline)
}
\keyword{datasets}
