% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nycc_ct_20}
\alias{nycc_ct_20}
\title{2020 Census Tract Shapefile (clipped to shoreline)}
\format{
## `nycc_ct_20`
An sf collection with 2325 rows and 17 columns w/ multiploygon geography:
}
\source{
<https://s-media.nyc.gov/agencies/dcp/assets/files/zip/data-tools/bytes/nyct2010_23a.zip>
}
\usage{
nycc_ct_20
}
\description{
2020 Census Tract Shapefile (clipped to shoreline)
}
\keyword{datasets}
