% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pretty_date}
\alias{pretty_date}
\title{Make a pretty date}
\usage{
pretty_date(x = NULL)
}
\arguments{
\item{x}{A \code{Date}. If \code{x} is \code{NULL}, the system date is returned.}
}
\value{
A character string representing the date.
}
\description{
Make a pretty date
}
\examples{

pretty_date()

}
