% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_nycc}
\alias{scale_fill_nycc}
\title{Color and fill scales for ggplots}
\usage{
scale_fill_nycc(palette = "mixed", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{One of \code{"mixed", "cool", "warm", "diverging"}}

\item{discrete}{Boolean, should the scale be discrete?}

\item{reverse}{Boolean, reverse the order of the selected palette}

\item{...}{Further arguments passed to \code{scale_*} from \code{ggplot2}}
}
\description{
Color and fill scales for ggplots
}
