% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{unzip_sf}
\alias{unzip_sf}
\title{Helper function for loading in shapefiles that come zipped}
\usage{
unzip_sf(zip_url)
}
\arguments{
\item{zip_url}{The url that the file you need to download and unzip is at}
}
\value{
The temporary path the file is stored in, ready to pass to st_read() or other
}
\description{
Helper function for loading in shapefiles that come zipped
}
