% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/council.R
\name{council_pdf}
\alias{council_pdf}
\title{Title}
\usage{
council_pdf(toc = FALSE, highlight = "default", ...)
}
\arguments{
\item{toc}{Whether to include a table of contents. Defaults to \code{FALSE}.}

\item{highlight}{Highlighting style}

\item{...}{Further options passed to \code{bookdown::pdf_document2()}.}
}
\description{
Title
}
