% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapshot.R
\name{mapshot}
\alias{mapshot}
\title{mapshot pdf}
\usage{
mapshot(
  x,
  url = NULL,
  file = NULL,
  remove_controls = c("zoomControl", "layersControl", "homeButton", "scaleBar",
    "drawToolbar", "easyButton"),
  ...
)
}
\description{
This is a \code{mapshot} adjustment that saves maps as pdf and
warns users if they choose png
}
\examples{
\dontrun{
library(leaflet)

m <- leaflet() \%>\%
 addTiles()
mapshot(m, file = "test.png")
file.remove("test.pdf")
}

}
