% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{nycc_pal}
\alias{nycc_pal}
\title{Make a color palette with NYCC colors}
\usage{
nycc_pal(palette = "mixed", reverse = FALSE, ...)
}
\arguments{
\item{palette}{One of \code{"mixed", "cool", "warm", "diverging"}}

\item{reverse}{Boolean, reverse the order of the selected palette}

\item{...}{Further arguments passed to \code{colorRampPalette}}
}
\value{
A function made by \code{colorRampPalette}
}
\description{
Make a color palette with NYCC colors
}
