% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R, R/test.R
\name{pal_nycc}
\alias{pal_nycc}
\title{Make a color palette with NYCC colors. Second iteration from nycc_pal.}
\usage{
pal_nycc(palette = "main", reverse = FALSE, ...)

pal_nycc(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{One of \code{"bw","main", "mixed", "cool", "warm", "diverging", "indigo", "blue", "violet", "bronze", "orange", "forest"}}

\item{reverse}{Boolean, reverse the order of the selected palette}

\item{...}{Further arguments passed to \code{colorRampPalette}}
}
\value{
The palette inputted, forward or reverse, grabbed from nycc_palettes and with additional palette options for \code{"single", "double"}

A function made by \code{colorRampPalette}
}
\description{
Make a color palette with NYCC colors. Second iteration from nycc_pal.

Make a color palette with NYCC colors. Second itertion from nycc_pal.
}
