% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pal_nycc}
\alias{pal_nycc}
\title{Make a color palette with NYCC colors. Second iteration from nycc_pal.}
\usage{
pal_nycc(palette = "main", reverse = FALSE)
}
\arguments{
\item{palette}{One of \code{"bw","main", "mixed", "nycc_blue", "cool", "warm", "diverging", "indigo", "blue", "violet", "bronze", "orange", "forest", "single", "double"}}

\item{reverse}{Boolean, reverse the order of the selected palette}
}
\value{
The palette inputted, forward or reverse, grabbed from nycc_palettes and with additional palette options for \code{"single", "double"}
}
\description{
Make a color palette with NYCC colors. Second iteration from nycc_pal.
}
