% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_nycc}
\alias{scale_fill_nycc}
\alias{scale_color_nycc}
\alias{scale_color_discrete}
\alias{scale_color_continuous}
\alias{scale_fill_discrete}
\alias{scale_fill_continuous}
\title{Color and fill scales for ggplots}
\usage{
scale_fill_nycc(palette = "mixed", discrete = TRUE, reverse = FALSE, ...)

scale_color_nycc(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_color_discrete(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_color_continuous(...)

scale_fill_discrete(palette = "mixed", discrete = TRUE, reverse = FALSE, ...)

scale_fill_continuous(...)
}
\arguments{
\item{palette}{One of \code{"bw","main", "mixed", "cool", "warm", "diverging"}}

\item{discrete}{Boolean, should the scale be discrete?}

\item{reverse}{Boolean, reverse the order of the selected palette}

\item{...}{Further arguments passed to \code{scale_*} from \code{ggplot2}}
}
\description{
The functions \code{scale_*_continuous} and \code{scale_*_discrete} are
exported from this package as aliases for the functions \code{scale_*_nycc}
with appropriate default arguments. Because of this, these colors will
overwrite \code{ggplot2}'s default scales. To prevent this, either set scales
manually in plots by calling \code{ggplot2::scale_*}, or attach
\code{ggplot2} after \code{councildown}.
}
\details{
When \code{discrete} is \code{TRUE} arguments are passed via \code{...} to
\code{\link[ggplot2]{discrete_scale}}. This is the default behavior.
Otherwise, \code{...} arguments are passed to
\code{\link[ggplot2]{scale_color_gradientn}} or
\code{\link[ggplot2]{scale_fill_gradientn}} as appropriate.
}
